+++ ChannelIO/Source/Models/CHVeil.swift
@@ -25,6 +25,7 @@ struct CHVeil: CHGuest, CHEntity {
   
   var country: String = ""
   var city: String = ""
+  var locale: String = ""
   
   var createdAt: Date?
   var updatedAt: Date?
@@ -56,6 +57,7 @@ extension CHVeil: Mappable {
     profile         <- map["profile"]
     country         <- map["country"]
     city            <- map["city"]
+    locale          <- map["locale"]
     createdAt       <- (map["createdAt"], CustomDateTransform())
     updatedAt       <- (map["updatedAt"], CustomDateTransform())
   }