+++ ChannelIO/Source/Services/RestService.swift
@@ -141,9 +141,9 @@ enum RestRouter: URLRequestConvertible {
     case .UnregisterToken(let key, _):
       return "/app/device_tokens/ios/\(key)"
     case .SendEvent(let pluginId, _):
-      return "/app/plugins/\(pluginId)/events"
+      return "/app/plugins/\(pluginId)/events/v2"
     case .GetCountryCodes:
-      return "/countries"
+      return "/available/countries"
     case .GetFollowingManager:
       return "/app/channels/following_managers"
     case .RequestProfileBot(let pluginId, let chatId):
@@ -184,6 +184,7 @@ enum RestRouter: URLRequestConvertible {
     }
     
     headers["Accept"] = "application/json"
+    headers["User-Agent"] = CHUtils.generateUserAgent()
     
     let now = Date()
     let cookies = HTTPCookieStorage.shared.cookies?