+++ ChannelIO/Source/Utils/CHUtils.swift
@@ -207,6 +207,25 @@ class CHUtils {
       .infoDictionary?["CFBundleShortVersionString"] as? String ?? "unknown"
     return params
   }
+  
+  class func generateUserAgent() -> String {
+    let version = Bundle(for: ChannelIO.self)
+      .infoDictionary?["CFBundleShortVersionString"] as? String ?? "unknown"
+    
+    let deviceType = UIDevice.current.model.hasPrefix("iPad") ? "iPad" : "iPhone"
+    var ua = "Mozilla/5.0"
+    ua += " (\(deviceType);"
+    ua += " CPU"
+    ua += deviceType == "iPhone" ? " iPhone" : ""
+    ua += " OS \(UIDevice.current.systemVersion.replace(".", withString: "_"))"
+    //iphone or ipad
+    //os version
+    ua += " like Mac OS X)"
+    ua += " Mobile"
+    ua += " ChannelSDK/\(version)"
+    //Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_2 like Mac OS X) Mobile ChannelPlugin/1.0.0
+    return ua
+  }
 }
 
 typealias dispatchClosure = () -> Void