+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 5.2.0
+## Updates
+* Updated pushbot flow logics
+* Added pushbot button and image redirection
+* Added keep push bot button
+* Updated event api 
+* Removed target evaluation logic
+
 # 5.1.8
 ## Bug fixes
 * Fixed int overflow issue for requestId